requirements.txt
# requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.8.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dateutil==2.9.0
pytest==8.3.4
httpx==0.28.1
cryptography==41.0.7
APScheduler==3.10.4